[TEST] Reorganize and consolidate all unflatten mode tests into tests/test_unflatten_comprehensive.py
